src/requirements.txt
torch
torchvision
torchsummary
matplotlib
numpy
pandas
scikit-learn
tqdm
Pillow
opencv-python
